+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   fibers even more. This does not affect having to allocate memory using mmap for the actual stack.
   This should also reduce memory use in the long term as this can reuse memory much more efficiently than the D
   conservative GC can free it.
+- Reduce allocations in the reactor
 
 ## `0.0.1-beta.1`